fix type size computations in parser
The type sizes are important to get right because
the ABI relies on them when it emits memory blits
to pass/return structs.

+++ all.h
@@ -431,7 +431,7 @@ struct Fn {
 struct Typ {
 	char name[NString];
 	int dark;
-	uint size;
+	ulong size;
 	int align;
 
 	struct {
